+++ client/src/components/game/mechanics/eventListeners/addDirectionDetection.js
