+++ client/src/components/game/mechanics/eventListeners/addVisibilityDetection.test.js
